fixed a few bugs. added student removal feature
fixed a bug where if a teacher presses on a student in the student list the app would crash.
fixed a bug where if a student does not have a teacher and goes into my teacher he would get the wrong error.
added the ability for a teacher to remove a student